requirements.txt
asttokens==2.0.5
backcall==0.2.0
black==21.12b0
click==8.0.3
decorator==5.1.1
elliptic-curve==0.1.2
executing==0.8.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
ipdb==0.13.9
ipython==8.1.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy==0.931
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-dotenv==0.19.2
python-http-client==3.3.7
sendgrid==6.4.0
six==1.16.0
SQLAlchemy==1.4.29
SQLAlchemy-Utils==0.38.2
stack-data==0.2.0
starkbank-ecdsa==2.0.3
toml==0.10.2
tomli==1.2.3
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5
Werkzeug==2.0.2
